Move ConsoleMonitor to core-spi
